Fix package name; add HAS_SPINNERS and HAS_BUTTONS constants to ObjectiveMatchFragment
