+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 32
 
     defaultConfig {
-        applicationId "org.stormrobotics.scoutingradar2022"
+        applicationId "org.stormroboticsnj.scoutingradar2022"
         minSdk 19
         targetSdk 32
         versionCode 1
