+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.stormrobotics.scoutingradar2022">
+    package="org.stormroboticsnj.scoutingradar2022">
 
     <application
         android:allowBackup="true"
@@ -10,7 +10,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.ScoutingRadar2022.Charge">
         <activity
-            android:name=".MainActivity"
+            android:name="org.stormroboticsnj.scoutingradar2022.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
