+++ app/src/main/java/org/stormroboticsnj/scoutingradar2022/SettingsFragment.java
@@ -1,0 +1,13 @@
+package org.stormroboticsnj.scoutingradar2022;
+
+import android.os.Bundle;
+
+import androidx.preference.PreferenceFragmentCompat;
+
+public class SettingsFragment extends PreferenceFragmentCompat {
+
+    @Override
+    public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
+        setPreferencesFromResource(R.xml.root_preferences, rootKey);
+    }
+}
