+++ app/src/main/res/layout/fragment_pit_scout.xml
@@ -3,7 +3,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".PitScoutFragment">
+        tools:context=".scoutingfragments.PitScoutFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
