+++ app/src/main/res/layout/fragment_subjective_match.xml
@@ -3,7 +3,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SubjectiveMatchFragment">
+        tools:context=".scoutingfragments.SubjectiveMatchFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
